src/requirements.txt
streamlit==1.41.1
scikit-learn>=1.1.0
google-generativeai
python-dotenv
pandas==2.2.3
numpy==2.2.2
plotly
